test_operate vault withdraw passed.
We should liquidate our token positions to want token In the liquidatePosition(). In addition, we should calculate and return the amount of liquidated assets as _liquidatedAmount. If you do not return the _liquidatedAmount, the Vault will not know how much has been liquidated, and will not transfer want token to  user.